Fixed getters of VectorTypes and TransformTypes to return references. Fixed an issue in the GraphicsManager where an inline function was not implemented in the header.

+++ src/engine/Game.cpp
@@ -110,7 +110,7 @@ void Engine::Game::RunVariableFramerate()
 		Draw(m_GameTime);
 
 		// Update the timing information
-		m_GameTime.deltaTimeMicros = std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::high_resolution_clock::now() - previousTime).count();
+		m_GameTime.deltaTimeMicros = (unsigned int)std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::high_resolution_clock::now() - previousTime).count();
 		m_GameTime.frameCount++;
 		m_GameTime.totalTimeMicros += m_GameTime.deltaTimeMicros;
 		previousTime = std::chrono::high_resolution_clock::now();
