+++ src/engine/input/MouseListener.hpp
@@ -20,7 +20,7 @@ namespace Engine{
 		virtual void ProcessMousePositionEvent(double xPos, double yPos) { };
 		virtual void ProcessMouseEnterLeaveEvent(bool entered) { };
 		virtual void ProcessMouseButtonEvent(int button, MouseButtonAction action) { };
-		virtual void ProcessMouseScrollEvent(int xOffset, int yOffset) { };
+		virtual void ProcessMouseScrollEvent(double xOffset, double yOffset) { };
 
 	};
 }
